feat: 에디터 이미지 붙여넣기 이벤트 핸들러 수정 - #64
